inline comparative stats helpers
